requirements.txt
Flask==0.10.1
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==1.3.1
Flask-SQLAlchemy==2.0
Flask-Script==2.0.5
Flask-WTF==0.11
Jinja2==2.7.3
Mako==1.0.1
MarkupSafe==0.23
SQLAlchemy==0.7.10
WTForms==2.0.2
Werkzeug==0.10.1
alembic==0.7.5.post2
blinker==1.3
ipython==2.2.0
itsdangerous==0.24
pillow==2.7.0
pyreadline==2.0
requests==1.1.0